Use host binds instead of tmpfs for EmptyDir
Starting from this change EmptyDir volumes will be created on the host and
mounted in the container using Docker's bind mounts. The reason for this
is greater alignement with the semantics of EmptyDir in K8s. This will
have 2 consequences:
- the volumes will no longer be deleted upon container stop (or crash),
- the volumes will be visible outside of the container.

This also means that there is no longer a need to special-case EmptyDirs
conlicting with tmpfs in the container, so the respective workaround was
removed.