Remove some Konlet systemd dependencies
Konlet service should not want the docker-events-collector.service,
because it should only be run if the relevant metadata is present.
Konlet also does not need the "after stackdriver-logging.service"
declaration, because it is ineffective due to how this service is
started, and it is not required (all logs will be picked up even if
stackdriver-logging service starts later in the process).